requirements.txt
# Flask and web framework dependencies
Flask==2.3.3
python-dotenv==1.0.0

# OpenAI integration
openai==1.3.5

# Data processing and numerical computation
numpy==1.24.3
pandas==2.0.3

# Image processing (for LUT operations)
Pillow==10.0.1
opencv-python==4.8.1.78

# HTTP requests
requests==2.31.0

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0

# CORS support for frontend integration
flask-cors==4.0.0

# JSON handling
jsonschema==4.19.1 